:rotating_light: Fixing mypy warnings

+++ src/python/katas/py7kyu/make_a_funciton_that_does_arithmetic.py
@@ -8,7 +8,7 @@ ops = {
 }
 
 
-def arithmetic(a: int, b: int, operator) -> int:
+def arithmetic(a: int, b: int, operator) -> float:
     return ops[operator](a, b)
